+++ packages/focusjail/src/FocusJailContainer.js
@@ -1,0 +1,21 @@
+/**
+ * Copyright Zendesk, Inc.
+ *
+ * Use of this source code is governed under the Apache License, Version 2.0
+ * found at http://www.apache.org/licenses/LICENSE-2.0.
+ */
+
+import PropTypes from 'prop-types';
+
+import { useFocusJail } from './useFocusJail';
+
+export function FocusJailContainer({ children, render = children, ...options }) {
+  return render(useFocusJail(options));
+}
+
+FocusJailContainer.propTypes = {
+  children: PropTypes.func,
+  render: PropTypes.func,
+  focusOnMount: PropTypes.bool,
+  environment: PropTypes.object
+};