+++ packages/selection/src/useSelection.js
@@ -106,7 +106,7 @@ function stateReducer(state, action, { focusedItem, selectedItem, onFocus, onSel
       if (onSelect) {
         onSelect(action.payload);
 
-        return state;
+        return { ...state };
       }
 
       return { ...state, selectedItem: action.payload };
@@ -195,8 +195,8 @@ export function useSelection({
       controlledFocusedItem === undefined ? isSelected : controlledFocusedItem === item;
     const tabIndex =
       isFocused ||
-      (state.selectedItem === undefined &&
-        state.focusedItem === undefined &&
+      (controlledSelectedItem === undefined &&
+        controlledFocusedItem === undefined &&
         items.indexOf(item) === defaultFocusedIndex)
         ? 0
         : -1;